remove old reports
