+++ Cargo.toml
@@ -41,7 +41,7 @@ siphasher = "1.0.1"
 thegraph-core = "0.4.1"
 thegraph-graphql-http = "0.2.1"
 thiserror = "1.0.59"
-tokio = { version = "1.37", features = [
+tokio = { version = "1.38.0", features = [
     "macros",
     "parking_lot",
     "rt-multi-thread",
