+++ gateway-framework/src/lib.rs
@@ -10,8 +10,9 @@ pub mod http;
 pub mod indexing;
 pub mod ip_blocker;
 pub mod json;
+pub mod logging;
+pub mod metrics;
 pub mod network;
-pub mod reporting;
 pub mod scalar;
 pub mod topology;
 pub mod utils;
