--- gateway-framework/src/reporting/mod.rs
@@ -1,7 +1,0 @@
-mod kafka;
-mod logging;
-mod metrics;
-
-pub use kafka::{EventHandlerFn, KafkaClient};
-pub use logging::{error_log, init, LoggingOptions, CLIENT_REQUEST_TARGET, INDEXER_REQUEST_TARGET};
-pub use metrics::{with_metric, METRICS};
