+++ graph-gateway/src/main.rs
@@ -32,16 +32,12 @@ use gateway_framework::{
     },
     indexing::Indexing,
     ip_blocker::IpBlocker,
-    json,
+    json, logging,
     network::{
         discovery::Status,
         exchange_rate,
         indexing_performance::{IndexingPerformance, Status as IndexingPerformanceStatus},
         network_subgraph,
-    },
-    reporting::{
-        self, EventHandlerFn, KafkaClient, LoggingOptions, CLIENT_REQUEST_TARGET,
-        INDEXER_REQUEST_TARGET,
     },
     scalar::{self, ReceiptSigner},
     topology::network::{Deployment, GraphNetwork},
@@ -52,7 +48,6 @@ use graph_gateway::{
     indexers,
     indexers::indexing,
     indexings_blocklist::{self, indexings_blocklist},
-    reports::{report_client_query, report_indexer_query},
     subgraph_studio,
 };
 use ordered_float::NotNan;
@@ -98,31 +93,7 @@ async fn main() {
 
     let config_repr = format!("{config:?}");
 
-    // Instantiate the Kafka client
-    let kafka_client: &'static KafkaClient = match KafkaClient::new(&config.kafka.into()) {
-        Ok(kafka_client) => Box::leak(Box::new(kafka_client)),
-        Err(kafka_client_err) => {
-            tracing::error!(%kafka_client_err);
-            return;
-        }
-    };
-
-    // Initialize logging
-    reporting::init(
-        kafka_client,
-        LoggingOptions {
-            executable_name: "graph-gateway".into(),
-            json: config.log_json,
-            event_handler: EventHandlerFn::new(|client, metadata, fields| {
-                match metadata.target() {
-                    CLIENT_REQUEST_TARGET => report_client_query(client, fields),
-                    INDEXER_REQUEST_TARGET => report_indexer_query(client, fields),
-                    _ => unreachable!("invalid event target for KafkaLayer"),
-                }
-            }),
-        },
-    );
-
+    logging::init("graph-gateway".into(), config.log_json);
     tracing::info!("gateway ID: {}", gateway_id);
     tracing::debug!(config = %config_repr);
 
@@ -241,7 +212,6 @@ async fn main() {
             client: http_client.clone(),
         },
         receipt_signer,
-        kafka_client,
         budgeter,
         l2_gateway: config.l2_gateway,
         chains: Box::leak(Box::new(Chains::new(config.chain_aliases))),
@@ -316,7 +286,7 @@ async fn main() {
                         .allow_methods([http::Method::OPTIONS, http::Method::POST]),
                 )
                 // Set up the query tracing span
-                .layer(RequestTracingLayer::new(config.graph_env_id.clone()))
+                .layer(RequestTracingLayer)
                 // Set the query ID on the request
                 .layer(SetRequestIdLayer::new(gateway_id))
                 // Handle legacy in-path auth, and convert it into a header
